refactor TodoListContainer into a custom hook instead
