+++ src/components/App/App.js
@@ -11,6 +11,7 @@ import About from '../About/About';
 import './App.css';
 import TodoDetailData from '../../containers/TodoDetailData/TodoDetailData';
 import UpdateTodoControl from '../../containers/UpdateTodoControl/UpdateTodoControl';
+import List from '../List/List';
 
 
 export default function App() {
@@ -20,7 +21,7 @@ export default function App() {
         <Header />
         <Switch>
           <Route exact path='/' component={Landing} />
-          <Route exact path='/todos' component={TodoList} />
+          <Route exact path='/todos' component={List} />
           <Route exact path='/about' component={About} />
           <Route exact path='/todo/:_id' component={TodoDetailData} />
           <Route exact path='/todo/update/:_id' component={UpdateTodoControl} />
