+++ src/components/List/List.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { useTodoList } from '../../hooks/todoListHook';
+
+const List = () => {
+  const { todoNodes, todoFormOrNot } = useTodoList();
+
+  return (
+    <div className='TodoList'>
+      { todoFormOrNot() }
+        
+        {todoNodes}
+        
+      </div>
+  )
+}
+
+export default List;
